% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectEdgeDryspots.R
\name{calculateSamplePosition}
\alias{calculateSamplePosition}
\title{Classify sample position relative to array boundaries}
\usage{
calculateSamplePosition(spe, samples)
}
\arguments{
\item{spe}{A SpatialExperiment object}

\item{samples}{Character string specifying the sample ID column name}
}
\value{
Character vector of sample position classifications ("corner", "edge", "interior")
}
\description{
Classify sample position relative to array boundaries
}
\keyword{internal}
