% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_helper_functions.R
\name{problemAreas}
\alias{problemAreas}
\title{Enhanced problem areas detection with parameterized cluster size}
\usage{
problemAreas(
  .xyz,
  offTissue,
  uniqueIdentifier = NA,
  shifted = FALSE,
  min_cluster_size = 40
)
}
\arguments{
\item{.xyz}{Coordinate matrix with QC data}

\item{offTissue}{Vector of off-tissue spot names}

\item{uniqueIdentifier}{Sample identifier (default: NA)}

\item{shifted}{Whether to apply coordinate adjustment (default: FALSE)}

\item{min_cluster_size}{Minimum cluster size for morphological cleaning (default: 40)}
}
\value{
Data frame with problem area information
}
\description{
Enhanced problem areas detection with parameterized cluster size
}
